Merge, stage 1

+++ Commands/CmdDeleteLvl.cs
@@ -65,7 +65,7 @@ namespace MCForge
                     try { File.Delete("levels/level properties/" + message); }
                     catch { }
 
-                    MySQL.executeQuery("DROP TABLE `Block" + message + "`, `Portals" + message + "`, `Messages" + message + "`, `Zone" + message + "`");
+                    if (Server.useMySQL) MySQL.executeQuery("DROP TABLE `Block" + message + "`, `Portals" + message + "`, `Messages" + message + "`, `Zone" + message + "`"); else SQLite.executeQuery("DROP TABLE Block" + message + ", Portals" + message + ", Messages" + message + " Zone" + message + "");
 
                     Player.GlobalMessage("Level " + message + " was deleted.");
                 }
